Callback execution scheduled for the next tick
Checking for the conditions and executing the callbacks is now scheduled
for the next tick via setTimeout(0). This ensures that we are waiting for
the current function to return. Otherwise there could be async work
scheduled and then cancelled in the middle of execution, with further
async work scheduled before the function returns.
Closes I035254/visualtestjs#17